initialization structure implemented
